Update the stored procedure helpers in the micro benchmark
