+++ src/main/java/org/vanilladb/bench/server/param/micro/MicroTestbedLoaderSpParamHelper.java
@@ -15,35 +15,17 @@
  *******************************************************************************/
 package org.vanilladb.bench.server.param.micro;
 
+import org.vanilladb.bench.benchmarks.micro.rte.MicroTestbedParamHelper;
 import org.vanilladb.core.sql.Schema;
 import org.vanilladb.core.sql.storedprocedure.SpResultRecord;
 import org.vanilladb.core.sql.storedprocedure.StoredProcedureParamHelper;
 
-public class TestbedLoaderParamHelper extends StoredProcedureParamHelper {
-
-	private static final String TABLES_DDL[] = {
-			"CREATE TABLE item ( i_id INT, i_im_id INT, i_name VARCHAR(24), "
-					+ "i_price DOUBLE, i_data VARCHAR(50) )" };
-	private static final String INDEXES_DDL[] = {
-			"CREATE INDEX idx_item ON item (i_id)" };
-	
-	private int numOfItems = 0;
-
-	public String[] getTableSchemas() {
-		return TABLES_DDL;
-	}
-
-	public String[] getIndexSchemas() {
-		return INDEXES_DDL;
-	}
-	
-	public int getNumberOfItems() {
-		return numOfItems;
-	}
+public class MicroTestbedLoaderSpParamHelper extends MicroTestbedParamHelper
+		implements StoredProcedureParamHelper {
 
 	@Override
 	public void prepareParameters(Object... pars) {
-		numOfItems = (Integer) pars[0];
+		super.unpackParameters(pars);
 	}
 
 	@Override
@@ -56,4 +38,8 @@ public class TestbedLoaderParamHelper extends StoredProcedureParamHelper {
 		return new SpResultRecord();
 	}
 
+	@Override
+	public boolean isReadOnly() {
+		return false;
+	}
 }
