+++ src/main/java/org/vanilladb/bench/server/param/micro/MicroTxnSpParamHelper.java
@@ -15,6 +15,7 @@
  *******************************************************************************/
 package org.vanilladb.bench.server.param.micro;
 
+import org.vanilladb.bench.benchmarks.micro.rte.MicroTxnParamHelper;
 import org.vanilladb.core.sql.DoubleConstant;
 import org.vanilladb.core.sql.IntegerConstant;
 import org.vanilladb.core.sql.Schema;
@@ -23,35 +24,11 @@ import org.vanilladb.core.sql.VarcharConstant;
 import org.vanilladb.core.sql.storedprocedure.SpResultRecord;
 import org.vanilladb.core.sql.storedprocedure.StoredProcedureParamHelper;
 
-public class MicroTxnProcParamHelper extends StoredProcedureParamHelper {
+public class MicroTxnSpParamHelper extends MicroTxnParamHelper
+		implements StoredProcedureParamHelper {
 
-	private int readCount;
-	private int writeCount;
-	private int[] readItemId;
-	private int[] writeItemId;
-	private double[] newItemPrice;
 	private String[] itemName;
 	private double[] itemPrice;
-
-	public int getReadCount() {
-		return readCount;
-	}
-
-	public int getWriteCount() {
-		return writeCount;
-	}
-
-	public int getReadItemId(int index) {
-		return readItemId[index];
-	}
-
-	public int getWriteItemId(int index) {
-		return writeItemId[index];
-	}
-
-	public double getNewItemPrice(int index) {
-		return newItemPrice[index];
-	}
 
 	public void setItemName(String s, int idx) {
 		itemName[idx] = s;
@@ -63,30 +40,10 @@ public class MicroTxnProcParamHelper extends StoredProcedureParamHelper {
 
 	@Override
 	public void prepareParameters(Object... pars) {
-
-		// Show the contents of paramters
-	   //System.out.println("Params: " + Arrays.toString(pars));
-
-		int indexCnt = 0;
-
-		readCount = (Integer) pars[indexCnt++];
-		readItemId = new int[readCount];
-		itemName = new String[readCount];
-		itemPrice = new double[readCount];
-
-		for (int i = 0; i < readCount; i++)
-			readItemId[i] = (Integer) pars[indexCnt++];
-
-		writeCount = (Integer) pars[indexCnt++];
-		writeItemId = new int[writeCount];
-		for (int i = 0; i < writeCount; i++)
-			writeItemId[i] = (Integer) pars[indexCnt++];
-		newItemPrice = new double[writeCount];
-		for (int i = 0; i < writeCount; i++)
-			newItemPrice[i] = (Double) pars[indexCnt++];
-
-		if (writeCount == 0)
-			setReadOnly(true);
+		unpackParameters(pars);
+		
+		itemName = new String[getReadCount()];
+		itemPrice = new double[getReadCount()];
 	}
 
 	@Override
@@ -113,5 +70,4 @@ public class MicroTxnProcParamHelper extends StoredProcedureParamHelper {
 		}
 		return rec;
 	}
-
 }
